// app/shared/model/user.model.ts
export class User {
    ExternalMemberId: string;
    FirstName: string;
    LastName: string;
    DOB: string;
    Address1: string;
    State: string;
    Zip: string;
    Phone: string;
    Email: string;
    PlanId: string;
    PlanOption: string;
    Relationship: string;
    gender: string;
    PictureData: string;

}